added classifier to fingercounter
