+++ src/main.py
@@ -1,10 +1,11 @@
-from finger_counter_tester import FingerCounterTester
+from finger_counter2 import FingerCounter2
+from finger_counter3 import FingerCounter3
 
 if __name__ == "__main__":
     
     in_dir = "../test-data/"
-    video_name = "IMG_0204.MOV"
+    video_name = "IMG_0112.MOV"
     out_dir = "../test-data/out/"
     report_name = "report.txt"
-    fc = FingerCounterTester(video_name, report_name, in_dir, out_dir, True)
+    fc = FingerCounter3(video_name, report_name, in_dir, out_dir, True)
     fc.play_game()
