+++ src/shaker2.py
@@ -27,7 +27,7 @@ class Shaker():
 		num_min = 0
 		num_max = 0
 		start_amplitude = 10
-		amplitude = 20
+		amplitude = 15
 		mini = 9999
 		maxi = -9999
 		find = 'start' # find min/max
